Fix deploy

+++ scripts/deploy.sh
@@ -50,7 +50,9 @@ mv dist branches/$SOURCE_BRANCH
 shopt -s extglob
 rm -rf !(test-results|dist|assets|branches)
 
-rm .gitignore # remove our .gitignore so we can grab certain items
+if [ -f ".gitignore"]; then
+    rm .gitignore # remove our .gitignore so we can grab certain items
+fi
 
 # Move test-results output into source branch
 mv test-results branches/$SOURCE_BRANCH/unit-tests
